experimentform/requirements.txt
djangorestframework==3.12.4   
django-cors-headers==3.7.0    
django-environ==0.4.5    
Django==3.2.12   
pip==21.1
psycopg2-binary==2.9.1    
requests==2.26.0    